Split webui and API views in multiple files
