+++ poli/views/apiview.py
@@ -1,19 +1,14 @@
 """
     This file is part of Polichombr.
 
-    (c) 2016 ANSSI-FR
+    (c) 2018 ANSSI-FR
 
 
     Description:
         Routes for REST API
 """
 
-import os
-import datetime
-
 from poli import api, apiview, app
-from poli.models.family import FamilySchema
-from poli.models.sample import Sample, SampleSchema
 from poli.models.sample import FunctionInfoSchema
 from poli.models.yara_rule import YaraSchema
 from poli.models.models import TLPLevel
@@ -21,7 +16,11 @@ from poli.models.user import User
 
 from flask_security import auth_token_required
 
-from flask import jsonify, request, send_file, abort, make_response
+from flask import jsonify, request, abort, make_response
+
+from poli.views.api_family import *
+from poli.views.api_idaactions import *
+from poli.views.api_sample import *
 
 
 def plain_text(data):
@@ -86,6 +85,9 @@ def api_help():
 
 @apiview.route('/get_auth_token/', methods=["POST"])
 def generate_token():
+    """
+        Generate a temporary token for using the API
+    """
     data = request.json
     if not data:
         abort(400, "Missing JSON arguments")
@@ -96,679 +98,8 @@ def generate_token():
     return jsonify({'token': user.get_auth_token()})
 
 
-@apiview.route(
-    '/family/<family_id>/export/<tlp_level>/detection/yara',
-    methods=['GET'])
-def api_family_export_detection_yara(family_id, tlp_level):
-    """
-        This endpoint is ugly, should replace with tlp in argument
-    """
-    my_family = api.get_elem_by_type("family", family_id)
-    return plain_text(
-        api.familycontrol.export_yara_ruleset(my_family, tlp_level))
-
-
-@apiview.route(
-    '/family/<family_id>/export/<tlp_level>/detection/snort',
-    methods=['GET'])
-def api_family_export_detection_snort(family_id, tlp_level):
-    my_family = api.get_elem_by_type("family", family_id)
-    return plain_text(
-        api.familycontrol.export_detection_snort(my_family, tlp_level))
-
-
-@apiview.route(
-    '/family/<family_id>/export/<tlp_level>/detection/openioc/',
-    methods=['GET'])
-def api_family_export_detection_openioc(family_id, tlp_level):
-    """
-        This endpoint format should be reimplemented
-    """
-    my_family = api.get_elem_by_type("family", family_id)
-    return plain_text(
-        api.familycontrol.export_detection_openioc(my_family, tlp_level))
-
-
-@apiview.route(
-    '/family/<family_id>/export/<tlp_level>/detection/custom_elements',
-    methods=['GET'])
-def api_family_export_detection_custom_elements(family_id, tlp_level):
-    my_family = api.get_elem_by_type("family", family_id)
-    return plain_text(
-        api.familycontrol.export_detection_custom(my_family, tlp_level))
-
-
-@apiview.route(
-    '/family/<family_id>/export/<tlp_level>/samplesarchive',
-    methods=['GET'])
-def api_family_export_sampleszip(family_id, tlp_level):
-    my_family = api.get_elem_by_type("family", family_id)
-    zpath = api.familycontrol.generate_samples_zip_file(my_family, tlp_level)
-    if zpath is None:
-        return ""
-    return send_file("../" + zpath, as_attachment=True,
-                     attachment_filename="export.tar.gz")
-
-
-@apiview.route(
-    '/family/<family_id>/export/<tlp_level>/samplesioc',
-    methods=['GET'])
-def api_family_export_samplesioc(family_id, tlp_level):
-    my_family = api.get_elem_by_type("family", family_id)
-    return plain_text(
-        api.familycontrol.export_samplesioc(my_family, tlp_level))
-
-
-@apiview.route('/families/', methods=['GET'])
-@auth_token_required
-def api_get_families():
-    """
-        Exports all the families
-    """
-    result = api.familycontrol.get_all_schema()
-    return jsonify(result)
-
-
-@apiview.route('/family/', methods=['POST'])
-def api_post_families():
-    """
-        Insert a new family
-        @return the created family id
-    """
-    data = request.json
-    if data is None:
-        abort(400, "Missing JSON arguments")
-    fname = data['name']
-    tlp_level = TLPLevel.TLPAMBER
-    try:
-        tlp_level = data['tlp_level']
-    except KeyError:
-        app.logger.warning("No TLP for family, default to AMBER")
-
-    pfam = None
-    try:
-        if data['parent']:
-            pfam = api.familycontrol.get_by_name(data['parent'])
-    except KeyError:
-        pass
-
-    fam = api.familycontrol.create(fname, parentfamily=pfam)
-    if fam is None:
-        fid = None
-    else:
-        api.familycontrol.set_tlp_level(fam, tlp_level, no_propagation=True)
-        fid = fam.id
-    return jsonify({'family': fid})
-
-
-@apiview.route('/family/<fname>/', methods=['GET'])
-def api_get_family(fname):
-    """
-        Get family data using it's name
-    """
-    fam = api.familycontrol.get_by_name(fname)
-    if fam is None:
-        return jsonify({"family": None})
-    fschema = FamilySchema()
-    data = fschema.dump(fam).data
-    return jsonify({"family": data})
-
-
-@apiview.route('/family/<int:fid>/', methods=['GET'])
-def api_get_family_by_id(fid):
-    """
-        Get family informations
-    """
-    fam = api.get_elem_by_type("family", fid)
-    schema = FamilySchema()
-    result = schema.dump(fam).data
-    return jsonify({"family": result})
-
-
-@apiview.route('/family/<int:fid>/abstract/', methods=['POST'])
-def api_set_family_abstract(fid):
-    """
-        @arg abstract: The family abstract
-    """
-    if request.json is None:
-        abort(400, "Missing JSON data")
-
-    try:
-        family = api.get_elem_by_type("family", fid)
-        abstract = request.json["abstract"]
-        result = api.familycontrol.set_abstract(family, abstract)
-        return jsonify({"result": result})
-
-    except KeyError:
-        abort(400, "Missing abstract data")
-
-
-@apiview.route('/family/<int:fid>/yaras/', methods=['POST'])
-def api_add_yara_to_family(fid):
-    """
-        Add a yara rule to a family
-    """
-    family = api.get_elem_by_type("family", fid)
-    try:
-        rule_name = request.json["rule_name"]
-        rule = api.yaracontrol.get_by_name(rule_name)
-        if rule is None:
-            raise KeyError
-        result = api.yaracontrol.add_to_family(family, rule)
-    except KeyError:
-        abort(400, "Unknown yara")
-    return jsonify({"result": result})
-
-
-@apiview.route('/family/<fam_name>', methods=['POST'])
-def api_post_family(fam_name):
-    """
-        TODO: Update a family from POST request
-    """
-    abort(404)
-
-
-@apiview.route('/family/<int:family_id>/attachment/<int:file_id>/')
-def download_family_file(family_id, file_id):
-    """
-    Family attachment download endpoint.
-    """
-    attachment = api.get_elem_by_type("family_file", file_id)
-    data_file = attachment.filepath
-    if not os.path.exists(data_file):
-        abort(404)
-    return send_file('../' + data_file,
-                     as_attachment=True,
-                     attachment_filename=os.path.basename(data_file))
-
-
-@apiview.route('/samples/<shash>/')
-def api_get_sample_id_from_hash(shash):
-    """
-        Useful for initialization of scripts, get the remote
-        sample ID when you known only the sample hash
-    """
-    if len(shash) == 32:
-        sample = Sample.query.filter_by(md5=shash).first()
-    elif len(shash) == 40:
-        sample = Sample.query.filter_by(sha1=shash).first()
-    elif len(shash) == 64:
-        sample = Sample.query.filter_by(sha256=shash).first()
-    else:
-        abort(400, "Invalid hash length")
-    if sample is not None:
-        return jsonify({'sample_id': sample.id})
-    return jsonify({'sample_id': None})
-
-
-@apiview.route('/samples/<int:sid>/download/')
-def api_get_sample_file(sid):
-    """
-        Return the sample binary
-    """
-    sample = api.get_elem_by_type("sample", sid)
-    data_file = sample.storage_file
-    return send_file('../' + data_file,
-                     as_attachment=True,
-                     attachment_filename=os.path.basename(data_file))
-
-
-@apiview.route('/samples/', methods=['GET'])
-def api_get_samples():
-    """
-        Returns all the samples
-    """
-    result = api.samplecontrol.schema_export_all()
-    data = jsonify({'samples': result})
-    return data
-
-
-@apiview.route('/samples/', methods=['POST'])
-def api_post_samples():
-    """
-    @description : Insert a new sample in database, launch analysis
-    @arg: string filename
-    @arg: binary data : the sample content
-    @return : the sample ID
-    """
-
-    mfile = request.files['file']
-    if not mfile:
-        abort(400, "You must provide a file object")
-
-    tlp_level = TLPLevel.TLPAMBER
-    try:
-        tlp_level = int(request.form["tlp_level"])
-    except KeyError:
-        app.logger.debug("Could not find the tlp_level key")
-
-    try:
-        orig_filename = request.form['filename']
-    except KeyError:
-        app.logger.debug("No filename provided")
-        orig_filename = ""
-
-    samples = api.dispatch_sample_creation(mfile, orig_filename)
-    if len(samples) == 0:
-        abort(500, "Cannot create sample")
-
-    if tlp_level not in range(1, 6):
-        app.logger.warning("Incorrect TLP level, defaulting to AMBER")
-        tlp_level = TLPLevel.TLPAMBER
-
-    for sample in samples:
-        result = api.samplecontrol.set_tlp_level(sample, tlp_level)
-        if result is False:
-            app.logger.warning(
-                "Cannot set TLP level for sample %d" % sample.id)
-    result = api.samplecontrol.schema_export_many(samples)
-
-    return jsonify({'sample': result})
-
-
-@apiview.route('/samples/<int:sid>/', methods=['GET'])
-def api_get_unique_sample(sid):
-    sample_schema = SampleSchema()
-    data = Sample.query.get(sid)
-    if data is None:
-        return '{}'
-    result = sample_schema.dump(data).data
-    data = jsonify({'samples': result})
-    return data
-
-
-@apiview.route('/samples/<int:sid>/', methods=['POST'])
-def api_post_unique_sample(sid):
-    abort(405)
-
-
-@apiview.route('/samples/<int:sid>/analysis/', methods=['GET'])
-def api_get_sample_full_analysis(sid):
-    return jsonify({'analysis': 'Not implemented'})
-
-
-@apiview.route('/samples/<int:sid>/analysis/analyzeit/', methods=['GET'])
-def api_get_sample_analyzeit(sid):
-    return jsonify({'analyzeit': 'Not implemented'})
-
-
-@apiview.route('/samples/<int:sid>/analysis/strings/', methods=['GET'])
-def api_get_sample_strings(sid):
-    return jsonify({'strings': 'Not implemented'})
-
-
-@apiview.route('/samples/<int:sid>/analysis/peinfo/', methods=['GET'])
-def api_get_sample_peinfo(sid):
-    return jsonify({'peinfo': 'not implemented'})
-
-
-@apiview.route('/samples/<int:sid>/families/', methods=['POST'])
-def api_post_sample_family(sid):
-    samp = api.get_elem_by_type("sample", sid)
-    if request.json is None:
-        abort(400, "JSON not provided")
-    fam = None
-    if "family_id" in request.json.keys():
-        fid = request.json['family_id']
-        fam = api.get_elem_by_type("family", fid)
-    elif "family_name" in request.json.keys():
-        fname = request.json['family_name']
-        fam = api.familycontrol.get_by_name(fname)
-    else:
-        return jsonify({'result': False})
-    result = api.familycontrol.add_sample(samp, fam)
-
-    return jsonify({'result': result})
-
-
-@apiview.route('/samples/<int:sid>/abstract/', methods=['POST'])
-def api_set_sample_abstract(sid):
-    """
-        @arg: abstract Markdown for the abstract
-    """
-    data = request.json
-    if data is None or 'abstract' not in data.keys():
-        abort(400, 'Invalid JSON data provided')
-    abstract = data['abstract']
-    samp = api.get_elem_by_type("sample", sid)
-    result = api.samplecontrol.set_abstract(samp, abstract)
-    return jsonify({'result': result})
-
-
-@apiview.route('/samples/<int:sid>/abstract/', methods=['GET'])
-def api_get_sample_abstract(sid):
-    """
-        Returns the raw markdown sample abstract
-    """
-    sample = api.get_elem_by_type("sample", sid)
-    result = sample.abstract
-    return jsonify({'abstract': result})
-
-
-def get_filter_arguments(mrequest):
-    """
-        Get timestamp and address from request
-    """
-    data = mrequest.args
-    cur_timestamp, addr = None, None
-    if data is not None:
-        if 'timestamp' in data.keys():
-            cur_timestamp = data['timestamp']
-            form = "%Y-%m-%dT%H:%M:%S.%f"
-            try:
-                cur_timestamp = datetime.datetime.strptime(cur_timestamp, form)
-            except ValueError:
-                abort(500, "Wrong timestamp format")
-        if 'addr' in data.keys():
-            addr = int(data['addr'], 16)
-    return cur_timestamp, addr
-
-
-@apiview.route('/samples/<int:sid>/idaactions/', methods=['GET'])
-def api_get_idaactions_updates(sid):
-    """
-        Get all actions for a sample
-    """
-    timestamp = datetime.datetime.now()
-
-    actions = api.idacontrol.get_all(sid=sid, timestamp=timestamp)
-
-    form = "%Y-%m-%dT%H:%M:%S.%f"
-    str_time = datetime.datetime.strftime(timestamp, form)
-
-    return jsonify({'idaactions': actions,
-                    'timestamp': str_time})
-
-
-@apiview.route('/samples/<int:sid>/functions/', methods=['GET'])
-def api_get_sample_functions(sid):
-    functions = api.samplecontrol.get_functions(sid)
-    schema = FunctionInfoSchema(many=True)
-    return jsonify(schema.dump(functions).data)
-
-
-@apiview.route('/samples/<int:sid>/functions/proposednames/', methods=['GET'])
-def api_suggest_func_names(sid):
-    """
-        Returns a dictionary containing proposed function names
-        based on machoc matches.
-    """
-    sample = api.get_elem_by_type("sample", sid)
-    proposed_funcs = api.samplecontrol.get_proposed_funcnames(sample)
-    return jsonify({'functions': proposed_funcs})
-
-
-@apiview.route('/samples/<int:sid>/comments/', methods=['GET'])
-def api_get_sample_comments(sid):
-    """
-        Get all the comments for a given sample
-        @arg : address Get for one address
-                default : get all the comments
-        @arg : timestamp Limit the timeframe for comments
-                (ie, how old you want the comments)
-                default = 0, no limit
-    """
-    current_timestamp, addr = get_filter_arguments(request)
-    data = api.idacontrol.get_comments(sid, addr, current_timestamp)
-    return jsonify({'comments': data})
-
-
-@apiview.route('/samples/<int:sid>/comments/', methods=['POST'])
-def api_post_sample_comments(sid):
-    """
-        Upload a new comment for a sample
-    """
-    if request.json is None:
-        abort(400, "No JSON data")
-    data = request.json
-    if "address" not in data.keys() or "comment" not in data.keys():
-        abort(400, "Missing comment or address arguments")
-    address = data['address']
-    comment = data['comment']
-    app.logger.debug(
-        "Getting a new comment for sample %d : %s@0x%x",
-        sid,
-        comment,
-        address)
-    action_id = api.idacontrol.add_comment(address, comment)
-    result = api.samplecontrol.add_idaaction(sid, action_id)
-    return jsonify({'result': result})
-
-
-@apiview.route('/samples/<int:sid>/names/', methods=['GET'])
-def api_get_sample_names(sid):
-    """
-        Get names for a given sample
-        @arg : addr Get for one address
-                default : get all the names
-        @arg : timestamp Limit the timeframe for names
-                default = 0, no limit
-    """
-    current_timestamp, addr = get_filter_arguments(request)
-    data = api.idacontrol.get_names(sid, addr, current_timestamp)
-    return jsonify({'names': data})
-
-
-@apiview.route('/samples/<int:sid>/names/', methods=['POST'])
-def api_post_sample_names(sid):
-    """
-        Upload a new names for a sample
-        @arg addr the corresponding address
-        @arg name the name
-    """
-    data = request.json
-    addr = data['address']
-    name = data['name']
-    app.logger.debug(
-        "Getting a new name for sample %d : %s@0x%x",
-        sid,
-        name,
-        addr)
-    action_id = api.idacontrol.add_name(addr, name)
-    result = api.samplecontrol.add_idaaction(sid, action_id)
-    if result is True:
-        api.samplecontrol.rename_func_from_action(sid, addr, name)
-        # we don't care if the function is renamed for a global name,
-        # so if the name is created return True anyway
-    return jsonify({'result': result})
-
-
-@apiview.route('/samples/<int:sid>/types/', methods=['POST'])
-def api_post_sample_types(sid):
-    """
-        Manage the creation of type definitions at specific addresses
-    """
-    data = request.json
-    addr = data['address']
-    typedef = data['typedef']
-
-    action_id = api.idacontrol.add_typedef(addr, typedef)
-    result = api.samplecontrol.add_idaaction(sid, action_id)
-    return jsonify(dict(result=result))
-
-
-@apiview.route('/samples/<int:sid>/types/', methods=['GET'])
-def api_get_sample_types(sid):
-    """
-        Get the IDA types stored in DB
-    """
-    current_timestamp, addr = get_filter_arguments(request)
-    data = api.idacontrol.get_typedefs(sid, addr, current_timestamp)
-    return jsonify({'typedefs': data})
-
-
-@apiview.route('/samples/<int:sid>/structs/', methods=['POST'])
-def api_create_struct(sid):
-    """
-        Create a new IDA Struct for a given sample
-        @arg name: the structure name
-    """
-    data = request.json
-    if data is None:
-        abort(400, "Missing JSON data")
-    result = False
-    name = data['name']
-    app.logger.debug("Creating structure %s" % name)
-    mstruct = api.idacontrol.create_struct(name=name)
-    if mstruct is not False:
-        result = api.samplecontrol.add_idaaction(sid, mstruct)
-    return jsonify({'result': result, 'structs': [{'id': mstruct}]})
-
-
-@apiview.route('/samples/<int:sid>/structs/', methods=['GET'])
-def api_get_sample_structs(sid):
-    """
-        Returns the structures associated with a sample
-        @arg timestamp: get structs after this timestamp (optional)
-    """
-    timestamp = None
-    if request.args is not None and 'timestamp' in request.args.keys():
-        timestamp = request.args['timestamp']
-    structs = api.idacontrol.get_structs(sid, timestamp)
-    return jsonify({'structs': structs})
-
-
-@apiview.route('/samples/<int:sid>/structs/<int:struct_id>/', methods=['GET'])
-def api_get_one_structs(sid, struct_id):
-    structs = api.idacontrol.get_one_struct(struct_id)
-    return jsonify({'structs': structs})
-
-
-@apiview.route('/samples/<int:sid>/structs/<int:struct_id>/members/',
-               methods=['POST'])
-def api_create_struct_member(sid, struct_id):
-    """
-        Add a new member to a structure
-    """
-    result = False
-    data = request.json
-    if data is None:
-        abort(400, "Missing JSON data")
-    name = data["name"]
-    size = data["size"]
-    offset = data["offset"]
-    mid = api.idacontrol.create_struct_member(name=name,
-                                              size=size,
-                                              offset=offset)
-    if mid is None:
-        result = False
-    else:
-        result = api.idacontrol.add_member_to_struct(struct_id, mid)
-    return jsonify({'result': result})
-
-
-@apiview.route('/samples/<int:sid>/structs/<string:struct_name>/')
-def api_get_struct_by_name(sid, struct_name):
-    """
-        Get structure data from a name
-    """
-    result = api.idacontrol.get_struct_by_name(sid, struct_name)
-    return jsonify({'structs': result})
-
-
-@apiview.route('/samples/<int:sid>/structs/<int:struct_id>/',
-               methods=["PATCH"])
-def api_rename_struct(sid, struct_id):
-    """
-        Rename a struct
-    """
-    data = request.json
-    if data is None:
-        abort(400, "Missing JSON data")
-    name = data["name"]
-    result = api.idacontrol.rename_struct(struct_id, name)
-    return jsonify({'result': result})
-
-
-@apiview.route('/samples/<int:sid>/structs/<int:struct_id>/',
-               methods=["DELETE"])
-def api_delete_struct(sid, struct_id):
-    """
-        Completely delete a struct from database
-    """
-    result = api.idacontrol.delete_struct(struct_id)
-    return jsonify({"result": result})
-
-
-@apiview.route('/samples/<int:sid>/structs/<int:struct_id>/members/',
-               methods=['PATCH'])
-def api_update_struct_member(sid, struct_id):
-    data = request.json
-    if data is None:
-        abort(400, "Missing JSON data")
-    mid = data["mid"]
-    result = False
-    if 'newname' in data.keys():
-        result = api.idacontrol.change_struct_member_name(struct_id, mid,
-                                                          data["newname"])
-    if 'newsize' in data.keys():
-        result = api.idacontrol.change_struct_member_size(struct_id, mid,
-                                                          data["newsize"])
-    return jsonify({'result': result})
-
-
-@apiview.route('/samples/<int:sid>/structs/<int:struct_id>/members/',
-               methods=['GET'])
-def api_get_struct_member(sid, struct_id):
-    result = False
-    structs = None
-    return jsonify({'result': result, 'structs': structs})
-
-
-@apiview.route('/samples/<int:sid>/structs/<int:struct_id>/members/',
-               methods=['DELETE'])
-def api_delete_struct_member(sid, struct_id):
-    """
-        TODO : implement and test
-    """
-    result = False
-    return jsonify({'result': result})
-
-
-@apiview.route('/samples/<int:sid>/matches/', methods=['GET'])
-def api_get_matches(sid):
-    """
-    TODO
-        Get all the matches :
-            - Yara
-            - IAT hash
-            - Machoc
-    """
-    result = None
-    return jsonify({'result': result})
-
-
-@apiview.route('/samples/<int:sid>/matches/machoc', methods=['GET'])
-def api_get_machoc_matches(sid):
-    """
-        TODO : Get machoc hashes
-    """
-    sample = api.get_elem_by_type("sample", sid)
-    result = None
-    return jsonify({'result': result})
-
-
-@apiview.route('/samples/<int:sid>/matches/iat_hash', methods=['GET'])
-def api_get_iat_matches(sid):
-    """
-        TODO : Get IAT hashes
-    """
-    sample = api.get_elem_by_type("sample", sid)
-    result = None
-    return jsonify({'result': result})
-
-
-@apiview.route('/samples/<int:sid>/matches/yara', methods=['GET'])
-def api_get_yara_matches(sid):
-    """
-        TODO : Get yara matches
-    """
-    sample = api.get_elem_by_type("sample", sid)
-    result = None
-    return jsonify({'result': result})
-
-
 @apiview.route('/yaras/', methods=['GET'])
+@auth_token_required
 def api_get_all_yaras():
     """
         Dump all the yaras
